+++ lib/app/models/hierarchy.rb
