+++ lib/monarchy/version.rb
@@ -1,4 +1,4 @@
 # frozen_string_literal: true
 module Monarchy
-  VERSION = '0.2.0'
+  VERSION = '0.4.0'
 end
